+++ test/device_tests.jl
@@ -519,7 +519,7 @@ const POINTWISE_BODIES = (
     :_internal_energy_point!, :_interp_point!,
     :_kappa_point!, :_max_into_point!, :_metric_src_cyl_point!,
     :_metric_src_sph_point!, :_mole_fraction_point!,
-    :_mu_beta_point!, :_no_slip_wall_point!, :_nscbc_inflow_point!,
+    :_mu_beta_point!, :_no_slip_flux_point!, :_no_slip_wall_point!, :_nscbc_inflow_point!,
     :_nscbc_outflow_point!, :_pair_backward_local_point!,
     :_pair_backward_remote_point!, :_pair_forward_local_point!,
     :_pair_forward_remote_point!, :_pair_select_point!,